Fixes #15644 - Move enable/skip_orchestration to Host::Managed
skip_orchestration and enable_orchestration were called in Host::Base
when importing facts. These two methods could not be called there
because they come from the Orchestration mixin which Host::Base does not
have. This blocked importing facts for a Host::Discovered or a Host::Base.

To test this, I've moved the tests that import the facts to a separate
test class just for Host::Base. Notice I've moved a few other tests
which also belong there. There are probably more, but I didn't want to
make this more complicated.

Lastly, Nic::Managed called lookup_value_match on save, so importing
facts didn't work for Host::Base as that method wasn't there either, so
I moved it. Same logic applies to 'setup_clone'.
